+++ packages/node/test/manual/release-health/single-session/caught-exception-errored-session.js
@@ -1,40 +1,40 @@
 const Sentry = require('../../../../dist');
-const { assertSessions, constructStrippedSessionObject, BaseDummyTransport } = require('../test-utils');
+const {
+  assertSessions,
+  constructStrippedSessionObject,
+  BaseDummyTransport,
+  validateSessionCountFunction,
+} = require('../test-utils');
 
-let sessionCounter = 0;
-process.on('exit', ()=> {
-  if (process.exitCode !== 1) {
-    console.log('SUCCESS: All application mode sessions were sent to node transport as expected');
-  }
-})
+const sessionCounts = {
+  sessionCounter: 0,
+  expectedSessions: 2,
+};
+
+validateSessionCountFunction(sessionCounts);
 
 class DummyTransport extends BaseDummyTransport {
   sendSession(session) {
-    sessionCounter++;
-    if (sessionCounter === 1) {
-      assertSessions(constructStrippedSessionObject(session),
-        {
-          init: true,
-          status: 'ok',
-          errors: 1,
-          release: '1.1'
-        }
-      )
+    sessionCounts.sessionCounter++;
+
+    if (sessionCounts.sessionCounter === 1) {
+      assertSessions(constructStrippedSessionObject(session), {
+        init: true,
+        status: 'ok',
+        errors: 1,
+        release: '1.1',
+      });
     }
-    else if (sessionCounter === 2) {
-      assertSessions(constructStrippedSessionObject(session),
-        {
-          init: false,
-          status: 'exited',
-          errors: 1,
-          release: '1.1'
-        }
-      )
+
+    if (sessionCounts.sessionCounter === 2) {
+      assertSessions(constructStrippedSessionObject(session), {
+        init: false,
+        status: 'exited',
+        errors: 1,
+        release: '1.1',
+      });
     }
-    else {
-      console.log('FAIL: Received way too many Sessions!');
-      process.exit(1);
-    }
+
     return super.sendSession(session);
   }
 }
@@ -43,7 +43,7 @@ Sentry.init({
   dsn: 'http://test@example.com/1337',
   release: '1.1',
   transport: DummyTransport,
-  autoSessionTracking: true
+  autoSessionTracking: true,
 });
 
 /**
@@ -54,8 +54,7 @@ Sentry.init({
  * `beforeExit` event which happens right before the process exits.
  */
 try {
-  throw new Error('hey there')
-}
-catch(e) {
+  throw new Error('hey there');
+} catch (e) {
   Sentry.captureException(e);
 }