+++ packages/node/test/manual/release-health/single-session/terminal-state-sessions-sent-once.js
@@ -17,14 +17,13 @@ class DummyTransport extends BaseDummyTransport {
   sendSession(session) {
     sessionCounts.sessionCounter++;
 
-    if (sessionCounts.sessionCounter === 1) {
-      assertSessions(constructStrippedSessionObject(session), {
-        init: true,
-        status: 'crashed',
-        errors: 1,
-        release: '1.1',
-      });
-    }
+    assertSessions(constructStrippedSessionObject(session), {
+      init: true,
+      status: 'crashed',
+      errors: 1,
+      release: '1.1',
+    });
+
     return super.sendSession(session);
   }
 }